functionized ArrayToRecord
ArrayToRecord.vee has function ArrayToRecord and examples of how the code is writen and instantiate instances of the function.
The goal is userselected temperature, port, band, write that specific data into those fields within the record.  Field names would have direct tie to the type of data the setup ment to capture.