Hashing files with VEE
Computinghash.vee is a built-in example on how to use the Vee hash function.  Hash an Excel file 'works' but the hash changes if the excel template is saved, even if saved with no changes.  and where the excel file is saved changed the hash.  Best use is to hash contents of a file, like the data section only.
 ZBB