double checked ID2587
The N-6457 has NF problems on band1 and 2 but not other bands. Therefore only the ENR tables for band1 and 2 are bad and the others are good. The unit we shipped that passed was retested and passed again the the same results. The new build failed again with the same failed results. Therefore the ENR program must be broken.   This is the logic of the customers I work for.  It's disheartening.